+++ src/java/org/malbino/orion/entities/Empleado.java
@@ -6,7 +6,6 @@
 package org.malbino.orion.entities;
 
 import java.io.Serializable;
-import javax.persistence.Column;
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 import javax.persistence.PrimaryKeyJoinColumn;