+++ src/java/org/malbino/orion/servlets/ComprobantePagoEstudiante.java
@@ -7,6 +7,7 @@ package org.malbino.orion.servlets;
 
 import com.itextpdf.text.BadElementException;
 import com.itextpdf.text.BaseColor;
+import com.itextpdf.text.Chunk;
 import com.itextpdf.text.Document;
 import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.Font;
@@ -40,8 +41,8 @@ import org.malbino.orion.util.Redondeo;
  *
  * @author tincho
  */
-@WebServlet(name = "ComprobantePago", urlPatterns = {"/pagos/ComprobantePago"})
-public class ComprobantePago extends HttpServlet {
+@WebServlet(name = "ComprobantePagoEstudiante", urlPatterns = {"/pagos/ComprobantePagoEstudiante"})
+public class ComprobantePagoEstudiante extends HttpServlet {
 
     private static final String CONTENIDO_PDF = "application/pdf";
 
@@ -93,10 +94,11 @@ public class ComprobantePago extends HttpServlet {
                 document.newPage();
 
                 document.add(comprobante(comprobante, estudiante));
+                document.add(cuenta(request, estudiante));
 
                 document.close();
             } catch (IOException | DocumentException ex) {
-                Logger.getLogger(ComprobantePago.class.getName()).log(Level.SEVERE, null, ex);
+                Logger.getLogger(ComprobantePagoEstudiante.class.getName()).log(Level.SEVERE, null, ex);
             }
 
         }
@@ -106,7 +108,7 @@ public class ComprobantePago extends HttpServlet {
         PdfPTable table = new PdfPTable(100);
 
         //cabecera
-        String realPath = getServletContext().getRealPath("/resources/uploads/" + comprobante.getCarrera().getCampus().getInstituto().getLogo());
+        String realPath = getServletContext().getRealPath("/resources/uploads/" + comprobante.getInscrito().getCarrera().getCampus().getInstituto().getLogo());
         Image image = Image.getInstance(realPath);
         image.scaleToFit(160, 80);
         image.setAlignment(Image.ALIGN_CENTER);
@@ -158,7 +160,7 @@ public class ComprobantePago extends HttpServlet {
         table.addCell(cell);
 
         //fila 3
-        cell = new PdfPCell(new Phrase(comprobante.getCarrera().getCampus().getSucursal(), NORMAL));
+        cell = new PdfPCell(new Phrase(comprobante.getInscrito().getCarrera().getCampus().getSucursal(), NORMAL));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(50);
         cell.setBorder(Rectangle.LEFT | Rectangle.RIGHT);
@@ -177,7 +179,7 @@ public class ComprobantePago extends HttpServlet {
         table.addCell(cell);
 
         //fila 4
-        cell = new PdfPCell(new Phrase(comprobante.getCarrera().getCampus().getDireccion(), NORMAL));
+        cell = new PdfPCell(new Phrase(comprobante.getInscrito().getCarrera().getCampus().getDireccion(), NORMAL));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(50);
         cell.setBorder(Rectangle.LEFT | Rectangle.RIGHT);
@@ -196,8 +198,8 @@ public class ComprobantePago extends HttpServlet {
         table.addCell(cell);
 
         //fial 5
-        if (comprobante.getCarrera().getCampus().getTelefono() != null) {
-            cell = new PdfPCell(new Phrase(comprobante.getCarrera().getCampus().getTelefono().toString(), NORMAL));
+        if (comprobante.getInscrito().getCarrera().getCampus().getTelefono() != null) {
+            cell = new PdfPCell(new Phrase(comprobante.getInscrito().getCarrera().getCampus().getTelefono().toString(), NORMAL));
         } else {
             cell = new PdfPCell(new Phrase(" ", NORMAL));
         }
@@ -219,7 +221,7 @@ public class ComprobantePago extends HttpServlet {
         table.addCell(cell);
 
         //fila 6
-        cell = new PdfPCell(new Phrase(comprobante.getCarrera().getCampus().getInstituto().getUbicacion(), NORMAL));
+        cell = new PdfPCell(new Phrase(comprobante.getInscrito().getCarrera().getCampus().getInstituto().getUbicacion(), NORMAL));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(50);
         cell.setBorder(Rectangle.LEFT | Rectangle.BOTTOM | Rectangle.RIGHT);
@@ -251,7 +253,7 @@ public class ComprobantePago extends HttpServlet {
         cell.setBorder(Rectangle.LEFT | Rectangle.TOP);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase(comprobante.getCarrera().toString(), NORMAL));
+        cell = new PdfPCell(new Phrase(comprobante.getInscrito().getCarrera().toString(), NORMAL));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(50);
         cell.setBorder(Rectangle.TOP);
@@ -263,7 +265,7 @@ public class ComprobantePago extends HttpServlet {
         cell.setBorder(Rectangle.TOP);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase(comprobante.getGestionAcademica().toString(), NORMAL));
+        cell = new PdfPCell(new Phrase(comprobante.getInscrito().getGestionAcademica().toString(), NORMAL));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(10);
         cell.setBorder(Rectangle.TOP | Rectangle.RIGHT);
@@ -276,7 +278,7 @@ public class ComprobantePago extends HttpServlet {
         cell.setBorder(Rectangle.LEFT | Rectangle.BOTTOM);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase(comprobante.getEstudiante().toString(), NORMAL));
+        cell = new PdfPCell(new Phrase(comprobante.getInscrito().getEstudiante().toString(), NORMAL));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(50);
         cell.setBorder(Rectangle.BOTTOM);
@@ -288,7 +290,7 @@ public class ComprobantePago extends HttpServlet {
         cell.setBorder(Rectangle.BOTTOM);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase(comprobante.getGestionAcademica().getRegimen().getNombre(), NORMAL));
+        cell = new PdfPCell(new Phrase(comprobante.getInscrito().getGestionAcademica().getRegimen().getNombre(), NORMAL));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(15);
         cell.setBorder(Rectangle.BOTTOM | Rectangle.RIGHT);
@@ -337,13 +339,13 @@ public class ComprobantePago extends HttpServlet {
                 cell.setBorder(Rectangle.LEFT);
                 table.addCell(cell);
 
-                cell = new PdfPCell(new Phrase(detalle.getCodigo(), NORMAL));
+                cell = new PdfPCell(new Phrase(detalle.getConcepto().getCodigo(), NORMAL));
                 cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
                 cell.setColspan(20);
                 cell.setBorder(PdfPCell.NO_BORDER);
                 table.addCell(cell);
 
-                cell = new PdfPCell(new Phrase(detalle.getNombre(), NORMAL));
+                cell = new PdfPCell(new Phrase(detalle.getConcepto().getNombre(), NORMAL));
                 cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
                 cell.setColspan(50);
                 cell.setBorder(PdfPCell.NO_BORDER);
@@ -445,4 +447,58 @@ public class ComprobantePago extends HttpServlet {
         return table;
     }
 
+    public PdfPTable cuenta(HttpServletRequest request, Estudiante estudiante) throws BadElementException, IOException {
+        PdfPTable table = new PdfPTable(100);
+
+        PdfPCell cell = new PdfPCell(new Phrase(" ", NEGRITA));
+        cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
+        cell.setColspan(100);
+        cell.setBorder(Rectangle.NO_BORDER);
+        cell.setFixedHeight(180f);
+        table.addCell(cell);
+
+        Phrase phrase = new Phrase();
+        phrase.add(new Chunk("INSTRUCCIONES DE USO", BOLDITALIC));
+        cell = new PdfPCell(phrase);
+        cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
+        cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
+        cell.setColspan(100);
+        cell.setBorder(Rectangle.NO_BORDER);
+        table.addCell(cell);
+
+        phrase = new Phrase();
+        phrase.add(new Chunk(" ", BOLDITALIC));
+        cell = new PdfPCell(phrase);
+        cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
+        cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
+        cell.setColspan(100);
+        cell.setBorder(Rectangle.NO_BORDER);
+        table.addCell(cell);
+
+        phrase = new Phrase();
+        phrase.add(new Chunk("Para ingresar al sistema:\n\n", ITALIC));
+        phrase.add(new Chunk("1) Abre un navegador web y dirigete a la siguiente dirección ", ITALIC));
+
+        String requestURL = request.getRequestURL().toString().replaceAll(request.getRequestURI(), "");
+        phrase.add(new Chunk(requestURL + "\n", BOLDITALIC));
+
+        phrase.add(new Chunk("2) Ingresa tu usuario y contraseña\n\n", ITALIC));
+        phrase.add(new Chunk("Usuario: ", BOLDITALIC));
+        phrase.add(new Chunk(estudiante.getUsuario() + "\n", ITALIC));
+        phrase.add(new Chunk("Contraseña: ", BOLDITALIC));
+        phrase.add(new Chunk(estudiante.getContrasenaSinEncriptar() + "\n\n", ITALIC));
+        phrase.add(new Chunk("3) Utiliza el menu para acceder a las opciones del sistema\n\n", ITALIC));
+        phrase.add(new Chunk(
+                "En el sistema pordras realizar tu Inscripción por Internet, revisar tu Historial Académico, "
+                + "revisar tu Historial Económico y actualizar tus datos personales.",
+                BOLDITALIC));
+        cell = new PdfPCell(phrase);
+        cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
+        cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
+        cell.setColspan(100);
+        cell.setBorder(Rectangle.NO_BORDER);
+        table.addCell(cell);
+
+        return table;
+    }
 }